apps/api/pyproject.toml
[project]
name = "opsmesh-api"
version = "0.1.0"
description = "OpsMesh API — incident intelligence backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email] (>=2.0)",
    "pydantic-settings>=2.0",
    "redis>=5.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0",
    "rq>=1.16",
    "psycopg2-binary>=2.9",
    "python-jose[cryptography] (>=3.3)",
    "passlib[bcrypt] (>=1.7)",
    "bcrypt (>=4.0,<4.1)",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
    "mypy>=1.13",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
